+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Text field helper text component
+- Form field component
 ## [0.1.1] - 2018-09-05
 ### Fixed
 - Components only became available with Laravel 5.4, so don't allow any versions below that